Removed Bundler gemspec requirement.
Discovered it wasn't possible to install gem dependencies for this
project when developing locally. The first commit (i.e.
b4cf18724edecfb43de2e294a71e7921beb1dcba), which introduced this
requirement, doesn't explain *why* Bundler was required. Due to Bundler
being included as part of the Ruby 2.6.0 distribution and having no
explicit references to Bundler within the code base it is safe to not
require it.