Update to Rubocop ~> 1
They've finally went 1.0, we should use it in the hopes that they will
be more stable with their lint naming scheme.

This singular code change is from the update to 2.5+ paired with the
update to Rubocop 1. It's more efficient and easier to read than the old
version so we're taking it.