Replace Travis CI with GitHub Actions
This applies some more of the development harness to CI as well. We're
now running the linters and an adjusted matrix of Ruby versions. As part
of this change, we have a new "ci" Bundler group that is a subset of the
development group for the purposes of running on CI.

We're not using the included Bundler caching from ruby/setup-ruby
because I don't want to download unnecessary gems for the different tasks.