Track Gemfile.lock in the repository
I used to hold the belief that libraries should not check the
Gemfile.lock file into source control because a developer might already
have versions of dependencies that satisfy the requirements for the
project. Thus, I would ignore the Gemfile.lock so a developer was free
to test on whichever version(s) they had installed.

When talking with Rafael França, he pointed out that this makes it
harder for developers to onboard onto a project because a combination of
dependencies might not satisfy the Bundler requirements.

After being bit by this once again after coming back to a project from a
hiatus, I decided to heed this advice. This change ensures that both CI
and new contributors have the required dependencies as soon as they run
the install command.