Allow comparison against a baseline
This change introduces a new comparison ordering strategy where the
initial report serves as the baseline against which the job compares the
other reports.

This is useful in cases where you're testing out different
implementations and want to see whether they are better or worse (and
the comparative increase or decrease) in an easy-to-decipher order.